fixed a bug with the select query

+++ src/jdbms/sql/datatypes/IntSQLType.java
@@ -2,13 +2,13 @@ package jdbms.sql.datatypes;
 
 import jdbms.sql.datatypes.util.SQLTypeFactory;
 
-public class IntSQLType extends SQLType<Integer> {
+public class IntSQLType extends SQLType<Long> {
 	static {
 		SQLTypeFactory.getInstance().registerType("INTEGER", IntSQLType.class);
 		SQLTypeFactory.getInstance().registerType("INT", IntSQLType.class);
 	}
 	public IntSQLType(String value) {
-		super(value == null ? null : Integer.parseInt(value));
+		super(value == null ? null : Long.parseLong(value));
 	}
 
 	@Override
